Conflicts:
	community/kernel/pom.xml
	community/server-api/LICENSES.txt
	community/server-api/NOTICE.txt
	community/server-examples/LICENSES.txt
	community/server-examples/NOTICE.txt
	community/server/src/test/java/org/neo4j/server/rest/streaming/StreamingBatchOperationDocIT.java
	enterprise/enterprise-performance-tests/LICENSES.txt
	enterprise/enterprise-performance-tests/NOTICE.txt

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/nioneo/store/NodeStore.java
@@ -315,7 +315,7 @@ public class NodeStore extends AbstractRecordStore<NodeRecord> implements Store
                 do
                 {
                     cursor.setOffset( offset );
-                    recordIsInUse = recordIsInUse( (long) cursor.getByte() );
+                    recordIsInUse = recordIsInUse( cursor.getByte() );
                 } while ( cursor.retry() );
             }
             return recordIsInUse;
